replace mapper by macro
