+++ crates/burn-core/src/module/param/mod.rs
@@ -3,7 +3,6 @@ mod constant;
 mod group;
 mod id;
 mod primitive;
-mod require_grad;
 mod running;
 mod sync_once_cell;
 mod tensor;
@@ -13,6 +12,5 @@ pub use base::*;
 pub use constant::*;
 pub use group::*;
 pub use id::*;
-pub(crate) use require_grad::*;
 pub use running::*;
 pub use visitor::*;
